rewriting seeds
