+++ database/seeds/05_events.js
@@ -22,7 +22,7 @@ user_ids.forEach((id, index) => {
 
 // insert records in batches of 100 otherwise sqlite will complain about too many variables
 exports.seed = async knex => {
-  await knex('events').truncate()
+  await knex('events').delete()
   for (let i = 0; i < events.length / 100; i++) {
     await knex('events').insert(events.slice(i * 100, (i + 1) * 100))
   }
